backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
alembic==1.13.1
celery==5.3.4
redis==5.0.1
openai>=1.40.0
requests==2.31.0
pyjwt==2.8.0
cryptography==42.0.5
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2